flask/requirements.txt
Flask
pandas
seaborn
matplotlib
pymongo
python-dotenv